make vs2008 vcproj, to x64, ImageMagick 7.0.8

+++ comps/misc/EditorView/axBaseView.cpp
@@ -2071,7 +2071,7 @@ BOOL CAxBaseView::PreTranslateMessage(MSG* pMsg)
 
 	if( pMsg->message == WM_KEYDOWN )
 	{
-		return _OnKeyDown( pMsg->wParam, LOWORD( pMsg->lParam ), HIWORD( pMsg->lParam ) );
+		return _OnKeyDown( (UINT)pMsg->wParam, LOWORD( pMsg->lParam ), HIWORD( pMsg->lParam ) );
 	}
